+++ .gitignore
@@ -1,2 +1,4 @@
-img/*
-vid/*
+asset/img/hipoberat.webp
+asset/img/prmdmakan.webp
+asset/img/smbg.webp
+
