+++ js/search.js
@@ -1,22 +1,47 @@
-//grab id
 var keywordlive = document.getElementById('keywordlive');
-var tombolnyari = document.getElementById('tombolnyari');
 var wadah = document.getElementById('wadah');
 
-//event
-keywordlive.addEventListener('keyup', function() {
+// Fungsi untuk inisialisasi event tab
+function inisialisasiTab() {
+    document.querySelectorAll('.tab-btns').forEach(btnGroup => {
+        const buttons = btnGroup.querySelectorAll('.tab-btn');
+        buttons.forEach(button => {
+            button.addEventListener('click', () => {
+                const targetId = button.dataset.tab;
 
-    //mulai ajax
+                buttons.forEach(btn => btn.classList.remove('active'));
+                button.classList.add('active');
+
+                const card = button.closest('.card');
+                const contents = card.querySelectorAll('.tab-content');
+
+                contents.forEach(content => {
+                    content.style.display = 'none';
+                });
+
+                const activeContent = card.querySelector(`#${targetId}`);
+                if (activeContent) {
+                    activeContent.style.display = 'block';
+                }
+            });
+        });
+    });
+}
+
+
+inisialisasiTab();
+
+// Event live search
+keywordlive.addEventListener('keyup', function() {
     var xhr = new XMLHttpRequest();
 
-    //ajax siap?
     xhr.onreadystatechange = function() {
-        if( xhr.readyState == 4 && xhr.status == 200 ) {
+        if (xhr.readyState == 4 && xhr.status == 200) {
             wadah.innerHTML = xhr.responseText;
+            inisialisasiTab(); // <- ini penting
         }
-    }
+    };
 
-    //eksekusi ajax
-    xhr.open('GET', 'search/dx.php?keywordlive=' + keywordlive.value, true);
+    xhr.open('GET', 'search/dx.php?keywordlive=' + encodeURIComponent(keywordlive.value), true);
     xhr.send();
 });
